backend/requirements.txt
websocket-client
asyncio
websockets